Move code around
* put mysql stuff all in its own file
* move types from definition files to where they are used
* remove the "plural" types
