--- logger.go
@@ -1,28 +1,0 @@
-package main
-
-import (
-	"log"
-	"net/http"
-	"time"
-)
-
-type jsonErr struct {
-	Code int    `json:"code"`
-	Text string `json:"text"`
-}
-
-func Logger(inner http.Handler, name string) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		start := time.Now()
-
-		inner.ServeHTTP(w, r)
-
-		log.Printf(
-			"%s\t%s\t%s\t%s",
-			r.Method,
-			r.RequestURI,
-			name,
-			time.Since(start),
-		)
-	})
-}
