+++ repo_mem.go
@@ -1,5 +1,3 @@
-// +build router
-
 package main
 
 import (
